pyproject.toml
[tool.poetry]
name = "compotes"
version = "2.0.0"
description = "Track debts & pools"
authors = ["Guilhem Saurel <guilhem.saurel@laas.fr>"]
license = "BSD-2-Clause"

[tool.poetry.dependencies]
python = "^3.10"
ndh = "^5.11.3"
dmdm = "^2.0.3"
django-tables2 = "^2.4.1"
yeouia = "^4.0.0"
django-filter = "^22.1"
ipython = "^8.0.1"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
coverage = "^5.5"
black = "^22.3.0"
flake8 = "^3.9.2"
pre-commit = "^2.11.1"
pydocstyle = "^6.0.0"
safety = "^2.2.0"

[tool.poetry.group.prod]
optional = true

[tool.poetry.group.prod.dependencies]
gunicorn = "^20.1.0"
psycopg2 = "^2.9.5"

[tool.isort]
profile = "black"
known_django = "django"
sections = "FUTURE,STDLIB,DJANGO,THIRDPARTY,FIRSTPARTY,LOCALFOLDER"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
